New pumping logic and preventing races
Implemented logic to prevent a potential race condition with the drink
queue size variable.  I also set up the pumping logic to turn on all the
pumps that still have remaining volume at the same time.  Also the
interupt is set up to go off with a frequency in line with with flow rate
of the pump.  So each call to the interupt represents 1mL of fluid.